refactor: to use term legacy import instead of nald
