+++ app/services/import/legacy-import/constants.js
@@ -1,0 +1,21 @@
+'use strict'
+
+/**
+ * NALD region prefix from import.NALD_ABS_LICENCES.AREP_EIUC_CODE will be mapped to one of the below regions
+ *
+ */
+const regions = {
+  AN: 'Anglian',
+  MD: 'Midlands',
+  NO: 'Northumbria',
+  NW: 'North West',
+  SO: 'Southern',
+  SW: 'South West (incl Wessex)',
+  TH: 'Thames',
+  WL: 'Wales',
+  YO: 'Yorkshire'
+}
+
+module.exports = {
+  regions
+}
