+++ app/services/import/licence-validator.service.js
@@ -1,0 +1,20 @@
+'use strict'
+
+/**
+ * Maps the import data to the desired format
+ * @module LicenceValidatorService
+ */
+
+/**
+ * Validates a licence is in the correct shape and format to perist in the databasse
+ *
+ * @param {} licence - The licence to validator
+ * @returns {} a licence is the licence is valid, other wise throw an error
+ */
+function go (licence) {
+  return licence
+}
+
+module.exports = {
+  go
+}
