+++ app/services/import/licence.service.js
@@ -2,44 +2,26 @@
 
 /**
  * Imports a licence from the imports tables into the views
- * @module ImportLicenceService
+ * @module LicenceService
  */
 
-const FetchImportLicenceService = require('./fetch-licence-from-import.service.js')
-const FetchImportLicenceVersionsService = require('./fetch-licence-versions-from-import.service.js')
-const ImportServiceLicenceMapper = require('./mappers/import-service/licence.mapper.js')
 const PersistLicenceService = require('./persist-licence.service.js')
+const LicenceValidatorService = require('./licence-validator.service.js')
 
 /**
- * Imports a licence from the imports tables into the views
+ * Validates and persists licence
  *
- * @param {string} licenceRef - The licence reference of the licence
+ * The licence should be mapped and in the coreect format before being saved
+ *
+ * @param {Object} licenceData - The mapped licence data from the data source
  * @returns {Promise<Object>} an object representing the `pageData` needed by the licence summary template.
  */
-async function go (licenceRef) {
-  try {
-    console.debug('Ref: ', licenceRef)
-    const licenceData = await FetchImportLicenceService.go(licenceRef)
+async function go (licenceData) {
+  const validatedLicence = LicenceValidatorService.go(licenceData)
 
-    console.log('Licence data', licenceData)
+  const savedLicence = await PersistLicenceService.go(validatedLicence)
 
-    const { ID: licenceId, FGAC_REGION_CODE: regionCode } = licenceData
-
-    const licenceVersionsData = await FetchImportLicenceVersionsService.go(licenceId, regionCode)
-
-    const mappedLicenceData = ImportServiceLicenceMapper.go(licenceData, licenceVersionsData)
-
-    console.log('Mapped licence data', mappedLicenceData)
-    // validate all here -
-
-    // persist
-    const savedLicence = await PersistLicenceService.go(mappedLicenceData)
-
-    console.log('Saved licence', savedLicence)
-  } catch (e) {
-    //  Should we do this in the controller and pass the message back to import service ?
-    console.error('Licence', licenceRef, 'failed to import: ', e)
-  }
+  console.log('Saved licence', savedLicence)
 }
 
 module.exports = {
