--- app/services/import/validators/licence.validator.js
@@ -1,9 +1,0 @@
-'use strict'
-
-function LicenceValidator (licence) {
-
-}
-
-module.exports = {
-  LicenceValidator
-}
